requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
fastapi==0.75.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2==2.9.3
py-automapper==1.0.4
pydantic==1.9.0
python-interface==1.6.1
pythondi==1.2.4
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
zipp==3.7.0
